Trigger on button release instead of button presses to avoid bugs
Bug that can happen with button press:
Say your game opens the keyboard whenever you release button_A. Therefore, when you exit the keyboard by pressing A with the cursor on the checkmark, the keyboard will close on button_A's press, then on A's release, the game will launch the keyboard again. It is now impossible to leave the keyboard.