Preserve original image palette when storing images
When an indexed image was using a non-default palette, its GMV and
BMP file would not be correct, as a default palette was assumed.

This is fixed here.

For this to work, the color index is also made non-static. This was
needed because the home menu resets the TFT palette to the default
at the start of its update loop. The original palette is stored in
a local variable (and restored) but not available elsewhere.